Delete and rename.
